# MOVE - MessageParser class
- MessageParser class 이동 ( rpc_service.cpp -> message_parser.hpp )
 - 다른 곳에서도 사용하기 위함.
